requirements.txt
pandas==2.0.3
openpyxl==3.1.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytest==7.4.0
pytest-cov==4.1.0
reportlab==4.0.4
